backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.9.0
pydantic-settings==2.5.0
scikit-learn==1.5.2
joblib==1.4.2
numpy==1.26.4
python-multipart==0.0.12
